newest specimen of 4H4D series
Objectives of version 11: Modify the function swift_alter so that the password is outputted when it has been successfully created and saved. (It does not need confirmation because all of the non-seed keywords have remained the same; they have already been confirmed earlier. Meanwhile, the three seed words are trivial to input; they just require copying the word that is given in the prompt.)
Also: at the end of 'reconstruct' mode there is now the option to enter swift_alter
mode.
Also in 'alter' mode: when a new, altered, password has been saved, it is
subsequently outputted; that way it is not necessary to reconstruct it all over again.
(This is similar to the edit to swift_alter mode.)
Also at the end of reconstruct and alter modes: a new function: optional logging of
passwords in a password construction log. (This is just a text file that can be accessed
with any text editor.) (This is for additional safety.)