latest program in the 4H16A series
Objective: Add an output and a log section to the end of
mode alter.
Also: updating help_intro, turning it into a help scroll.
Also: propagating an update of the inputting system from
FourHundredWords2FourDigits_11.py hither.
Also: add swift_alter function and call it at the conclusion of
reconstruct mode. (Not at the conclusion of alter mode. Why not?
swift_alter is called at the end of alter mode in 4H4D(11) because
creating a new (word-labyrinth that outputs a) four-digit password
is always risky because there is a liability for a 'digit crash'.
The chance of not having a digit crash is (10*9*8*7)/(10*10*10*10)
 = 50.4%.)
 Also: new NATO_alphanumber and NATO_spell_out functions for help
 with outputting alphanumberic passwords.