.4H16A(5): corrections:
(1) Function index_word_count was entirely missing even though it was expected; so it is being supplied. (2) L949: 'prefix_changed' flag reset, to fix a backsliding of 'prefix'. (3) Function NATO_spell_out is altered so that the password is spelled out in four consecutive lines of four Alphanumbers each. (4) Descriptions of functions were missing; so they are being supplied.